Adding a cc_binary main() program
This is the smallest hello world example you can conceive of. Just a
tiny main() that prints a message and a cc_binary() rule to build it.

The empty WORKSPACE file tells Bazel that this is a Bazel project. The
actual build targets are defined in BUILD files.

Build project with command `bazel build ...` or
`bazel build //:stellar`