Adding Rocket cc_library target
The cc_binary() rule makes an executable or a shared library. To
create smaller pieces of the dependency graph, we create intermediate
libraries with the cc_library() rule.

Let's make a Rocket and launch it!

For the stellar target to get access to the rocket target, it's
visibility needs to be set to public.

Depending targets like stellar can access all files listed in the
other targets hdrs list.