Add failing unit test with cc_test
We are doing some TDD, so let's add a test and just enough code to
make it compile.

Normally tests are created with GTest where the testing framework
provides the main() function for the test executable. Bazel just cares
about the exit code.

Tests (all of them) are run with: bazel test ...