Using genrule to create a tar.gz bundle
There are other types of Bazel rules as well, not just cc_binary() and
cc_library(). One useful type is genrule(). This rule allows a program
to run with given inputs to create some output. Can be used for Python
code generation, or as we do here, to run system tools.

Since the build output is placed in the bazel-out/k8-fastbuild/bin/
directory, we use --transform to make the tar file just have a bin/